refactor: Merge of Store/Update Requests for Persona model. minor fixes on PersonaController methods.
