+++ backendweb/app/Persona.php
@@ -18,6 +18,15 @@ class Persona extends Model
     ];
 
     /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [
+        'updated_at', 'created_at', 'id'
+    ];
+
+    /**
      * 1-n database relation with Registro.
      * @return hasMany
      */
